+++ src/Akeneo/Pim/Automation/DataQualityInsights/back/Infrastructure/Symfony/Controller/AddWordInDictionaryController.php
@@ -13,7 +13,7 @@ declare(strict_types=1);
 
 namespace Akeneo\Pim\Automation\DataQualityInsights\Infrastructure\Symfony\Controller;
 
-use Akeneo\Pim\Automation\DataQualityInsights\Application\CriteriaEvaluation\Consistency\SupportedLocaleChecker;
+use Akeneo\Pim\Automation\DataQualityInsights\Application\CriteriaEvaluation\Consistency\SupportedLocaleValidator;
 use Akeneo\Pim\Automation\DataQualityInsights\Application\FeatureFlag;
 use Akeneo\Pim\Automation\DataQualityInsights\Domain\Model\Write;
 use Akeneo\Pim\Automation\DataQualityInsights\Domain\Repository\TextCheckerDictionaryRepositoryInterface;
@@ -36,9 +36,9 @@ class AddWordInDictionaryController
     private $featureFlag;
 
     /**
-     * @var SupportedLocaleChecker
+     * @var SupportedLocaleValidator
      */
-    private $supportedLocaleChecker;
+    private $supportedLocaleValidator;
 
     /**
      * @var TextCheckerDictionaryRepositoryInterface
@@ -50,12 +50,12 @@ class AddWordInDictionaryController
 
     public function __construct(
         FeatureFlag $featureFlag,
-        SupportedLocaleChecker $supportedLocaleChecker,
+        SupportedLocaleValidator $supportedLocaleValidator,
         TextCheckerDictionaryRepositoryInterface $textCheckerDictionaryRepository,
         EventDispatcherInterface $eventDispatcher
     ) {
         $this->featureFlag = $featureFlag;
-        $this->supportedLocaleChecker = $supportedLocaleChecker;
+        $this->supportedLocaleValidator = $supportedLocaleValidator;
         $this->textCheckerDictionaryRepository = $textCheckerDictionaryRepository;
         $this->eventDispatcher = $eventDispatcher;
     }
@@ -67,7 +67,7 @@ class AddWordInDictionaryController
             $localeCode = new LocaleCode($request->request->get('locale'));
             $productId = new ProductId($request->request->getInt('product_id'));
 
-            if (!$this->supportedLocaleChecker->isSupported($localeCode)) {
+            if (!$this->supportedLocaleValidator->isSupported($localeCode)) {
                 throw new \InvalidArgumentException('Unable to process locales that are not handled by spellchecker');
             }