+++ src/Akeneo/Pim/Automation/DataQualityInsights/back/Infrastructure/Symfony/Controller/CheckTextController.php
@@ -13,7 +13,7 @@ declare(strict_types=1);
 
 namespace Akeneo\Pim\Automation\DataQualityInsights\Infrastructure\Symfony\Controller;
 
-use Akeneo\Pim\Automation\DataQualityInsights\Application\CriteriaEvaluation\Consistency\SupportedLocaleChecker;
+use Akeneo\Pim\Automation\DataQualityInsights\Application\CriteriaEvaluation\Consistency\SupportedLocaleValidator;
 use Akeneo\Pim\Automation\DataQualityInsights\Application\CriteriaEvaluation\Consistency\TextChecker;
 use Akeneo\Pim\Automation\DataQualityInsights\Application\FeatureFlag;
 use Akeneo\Pim\Automation\DataQualityInsights\Domain\ValueObject\LocaleCode;
@@ -31,16 +31,16 @@ class CheckTextController
 
     private $textChecker;
 
-    private $supportedLocaleChecker;
+    private $supportedLocaleValidator;
 
     /** @var LoggerInterface */
     private $logger;
 
-    public function __construct(FeatureFlag $featureFlag, TextChecker $textChecker, SupportedLocaleChecker $supportedLocaleChecker, LoggerInterface $logger)
+    public function __construct(FeatureFlag $featureFlag, TextChecker $textChecker, SupportedLocaleValidator $supportedLocaleValidator, LoggerInterface $logger)
     {
         $this->featureFlag = $featureFlag;
         $this->textChecker = $textChecker;
-        $this->supportedLocaleChecker = $supportedLocaleChecker;
+        $this->supportedLocaleValidator = $supportedLocaleValidator;
         $this->logger = $logger;
     }
 
@@ -54,7 +54,7 @@ class CheckTextController
         $localeCode = new LocaleCode($request->request->get('locale'));
 
         // @todo[DAPI-601] can we use a more appropriate response ?
-        if (empty($text) || !$this->supportedLocaleChecker->isSupported($localeCode)) {
+        if (empty($text) || !$this->supportedLocaleValidator->isSupported($localeCode)) {
             return new JsonResponse(null, Response::HTTP_NOT_FOUND);
         }