+++ src/Akeneo/Pim/Automation/DataQualityInsights/tests/back/Specification/Infrastructure/Consistency/TextChecker/AspellDictionarySpec.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 namespace Specification\Akeneo\Pim\Automation\DataQualityInsights\Infrastructure\Consistency\TextChecker;
 
 use Akeneo\Pim\Automation\DataQualityInsights\Application\Clock;
+use Akeneo\Pim\Automation\DataQualityInsights\Application\CriteriaEvaluation\Consistency\SupportedLocaleValidator;
 use Akeneo\Pim\Automation\DataQualityInsights\Domain\Exception\UnableToRetrieveDictionaryException;
 use Akeneo\Pim\Automation\DataQualityInsights\Domain\Model\Dictionary;
 use Akeneo\Pim\Automation\DataQualityInsights\Domain\ValueObject\LanguageCode;
@@ -23,9 +24,10 @@ final class AspellDictionarySpec extends ObjectBehavior
         Clock $clock,
         AspellDictionaryLocalFilesystemInterface $localFilesystemProvider,
         FilesystemInterface $localFilesystem,
-        FilesystemInterface $sharedFilesystem
+        FilesystemInterface $sharedFilesystem,
+        SupportedLocaleValidator $supportedLocaleValidator
     ) {
-        $this->beConstructedWith($mountManager, $clock, $localFilesystemProvider);
+        $this->beConstructedWith($mountManager, $clock, $localFilesystemProvider, $supportedLocaleValidator);
 
         $localFilesystemProvider->getFilesystem()->willReturn($localFilesystem);
         $localFilesystemProvider->getAbsoluteRootPath()->willReturn('/tmp');
@@ -58,8 +60,14 @@ final class AspellDictionarySpec extends ObjectBehavior
     public function it_gets_up_to_date_local_dictionary_relative_file_path(
         Clock $clock,
         FilesystemInterface $localFilesystem,
-        FilesystemInterface $sharedFilesystem
+        FilesystemInterface $sharedFilesystem,
+        SupportedLocaleValidator $supportedLocaleValidator
     ) {
+        $localeCode = new LocaleCode('en_US');
+        $languageCode = new LanguageCode('en');
+
+        $supportedLocaleValidator->extractLanguageCode($localeCode)->willReturn($languageCode);
+
         $sharedFilesystem->has('consistency/text_checker/aspell/custom-dictionary-en.pws')->willReturn(true);
 
         $localFilesystem->has('consistency/text_checker/aspell/custom-dictionary-en.pws')->willReturn(true);
@@ -71,15 +79,21 @@ final class AspellDictionarySpec extends ObjectBehavior
         $clock->fromTimestamp(1577462283)->willReturn($fileDate);
         $clock->getCurrentTime()->willReturn($now);
 
-        $this->getUpToDateSpellerDictionary(new LocaleCode('en_US'))
+        $this->getUpToDateSpellerDictionary($localeCode)
             ->shouldBeLike(new Speller\Dictionary('/tmp/consistency/text_checker/aspell/custom-dictionary-en.pws'));
     }
 
     public function it_gets_up_to_date_speller_dictionary_by_downloading_it_if_does_not_exists_locally(
         Clock $clock,
         FilesystemInterface $localFilesystem,
-        FilesystemInterface $sharedFilesystem
+        FilesystemInterface $sharedFilesystem,
+        SupportedLocaleValidator $supportedLocaleValidator
     ) {
+        $localeCode = new LocaleCode('en_US');
+        $languageCode = new LanguageCode('en');
+
+        $supportedLocaleValidator->extractLanguageCode($localeCode)->willReturn($languageCode);
+
         $localFilesystem->has('consistency/text_checker/aspell/custom-dictionary-en.pws')->willReturn(false, false, true);
         $sharedFilesystem->has('consistency/text_checker/aspell/custom-dictionary-en.pws')->willReturn(true);
 
@@ -98,24 +112,45 @@ final class AspellDictionarySpec extends ObjectBehavior
         $clock->fromTimestamp(1577462283)->willReturn($fileDate);
         $clock->getCurrentTime()->willReturn($now);
 
-        $this->getUpToDateSpellerDictionary(new LocaleCode('en_US'))
+        $this->getUpToDateSpellerDictionary($localeCode)
             ->shouldBeLike(new Speller\Dictionary('/tmp/consistency/text_checker/aspell/custom-dictionary-en.pws'));
     }
 
     public function it_return_null_if_there_is_no_dictionary_for_the_given_locale(
         FilesystemInterface $localFilesystem,
-        FilesystemInterface $sharedFilesystem
+        FilesystemInterface $sharedFilesystem,
+        SupportedLocaleValidator $supportedLocaleValidator
     ) {
+        $localeCode = new LocaleCode('en_US');
+        $languageCode = new LanguageCode('en');
+
+        $supportedLocaleValidator->extractLanguageCode($localeCode)->willReturn($languageCode);
         $localFilesystem->has('consistency/text_checker/aspell/custom-dictionary-en.pws')->willReturn(false);
         $sharedFilesystem->has('consistency/text_checker/aspell/custom-dictionary-en.pws')->willReturn(false);
 
         $this->getUpToDateSpellerDictionary(new LocaleCode('en_US'))->shouldReturn(null);
     }
 
+    public function it_return_null_if_there_is_no_supported_language_for_the_given_locale(
+        SupportedLocaleValidator $supportedLocaleValidator
+    ) {
+        $localeCode = new LocaleCode('en_US');
+
+        $supportedLocaleValidator->extractLanguageCode($localeCode)->willReturn(null);
+
+        $this->getUpToDateSpellerDictionary($localeCode)->shouldReturn(null);
+    }
+
     public function it_throws_exception_if_unable_to_download_dictionary(
         FilesystemInterface $localFilesystem,
-        FilesystemInterface $sharedFilesystem
+        FilesystemInterface $sharedFilesystem,
+        SupportedLocaleValidator $supportedLocaleValidator
     ) {
+        $localeCode = new LocaleCode('en_US');
+        $languageCode = new LanguageCode('en');
+
+        $supportedLocaleValidator->extractLanguageCode($localeCode)->willReturn($languageCode);
+
         $localFilesystem->has('consistency/text_checker/aspell/custom-dictionary-en.pws')->willReturn(false);
         $sharedFilesystem->has('consistency/text_checker/aspell/custom-dictionary-en.pws')->willReturn(true);
 
@@ -125,13 +160,19 @@ final class AspellDictionarySpec extends ObjectBehavior
             Argument::any()
         )->shouldNotBeCalled();
 
-        $this->shouldThrow(UnableToRetrieveDictionaryException::class)->during('getUpToDateSpellerDictionary', [new LocaleCode('en_US')]);
+        $this->shouldThrow(UnableToRetrieveDictionaryException::class)->during('getUpToDateSpellerDictionary', [$localeCode]);
     }
 
     public function it_gets_up_to_date_local_dictionary_relative_file_path_by_downloading_it_if_is_older_than_a_day(
         Clock $clock,
-        FilesystemInterface $localFilesystem
+        FilesystemInterface $localFilesystem,
+        SupportedLocaleValidator $supportedLocaleValidator
     ) {
+        $localeCode = new LocaleCode('en_US');
+        $languageCode = new LanguageCode('en');
+
+        $supportedLocaleValidator->extractLanguageCode($localeCode)->willReturn($languageCode);
+
         $localFilesystem->has('consistency/text_checker/aspell/custom-dictionary-en.pws')->willReturn(true);
 
         $localFilesystem->getTimestamp('consistency/text_checker/aspell/custom-dictionary-en.pws')->willReturn(1577462283);
@@ -144,15 +185,21 @@ final class AspellDictionarySpec extends ObjectBehavior
 
         $localFilesystem->putStream(Argument::any())->shouldNotBeCalled();
 
-        $this->getUpToDateSpellerDictionary(new LocaleCode('en_US'))
+        $this->getUpToDateSpellerDictionary($localeCode)
             ->shouldBeLike(new Speller\Dictionary('/tmp/consistency/text_checker/aspell/custom-dictionary-en.pws'));
     }
 
     public function it_gets_up_to_date_local_dictionary_relative_file_path_by_downloading_if_it_is_older_than_a_day_and_if_shared_one_is_newer(
         Clock $clock,
         FilesystemInterface $localFilesystem,
-        FilesystemInterface $sharedFilesystem
+        FilesystemInterface $sharedFilesystem,
+        SupportedLocaleValidator $supportedLocaleValidator
     ) {
+        $localeCode = new LocaleCode('en_US');
+        $languageCode = new LanguageCode('en');
+
+        $supportedLocaleValidator->extractLanguageCode($localeCode)->willReturn($languageCode);
+
         $localFilesystem->has('consistency/text_checker/aspell/custom-dictionary-en.pws')->willReturn(true);
         $sharedFilesystem->has('consistency/text_checker/aspell/custom-dictionary-en.pws')->willReturn(true);
 
@@ -172,7 +219,7 @@ final class AspellDictionarySpec extends ObjectBehavior
             $resource
         )->shouldBeCalled();
 
-        $this->getUpToDateSpellerDictionary(new LocaleCode('en_US'))
+        $this->getUpToDateSpellerDictionary($localeCode)
             ->shouldBeLike(new Speller\Dictionary('/tmp/consistency/text_checker/aspell/custom-dictionary-en.pws'));
     }