+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - DAPI-635: Add spellcheck on WYSIWG editors
 - DAPI-798: Allow spelling suggestions after a title formatter check
 - RUL-20: Rule engine - As Julia, I would like to copy values from/to different attribute types
+- DAPI-895: As Julia, I'd like spell-check to be available for Norwegian
 
 ## New features